db_final/requirements.txt
asgiref==3.3.1
certifi==2020.11.8
chardet==3.0.4
Django==3.1.4
django-cors-headers==3.5.0
djangorestframework==3.12.2
idna==2.10
mysqlclient==2.0.1
numpy==1.19.4
python-dateutil==2.8.1
pytz==2020.4
requests==2.25.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
